% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/version.R
\name{BFInfo}
\alias{BFInfo}
\title{Prints the version information for the BayesFactor package}
\usage{
BFInfo(print = TRUE)
}
\arguments{
\item{print}{if \code{TRUE}, print version information to the console}
}
\value{
\code{BFInfo} returns a character string containing the version and
 revision number of the package..
}
\description{
Prints the version, revision, and date information for the BayesFactor package
}
\details{
This function prints the version and revision information for the BayesFactor
package.
}
\author{
Richard D. Morey (\email{richarddmorey@gmail.com})
}
\keyword{misc}

