% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aaClasses.R, R/methods-BFprobability.R
\docType{class}
\name{BFprobability-class}
\alias{-,BFprobability,numeric-method}
\alias{/,BFprobability,numeric-method}
\alias{BFprobability-class}
\alias{[,BFprobability,index,missing,missing-method}
\title{General S4 class for representing multiple model probability comparisons}
\usage{
\S4method{/}{BFprobability,numeric}(e1, e2)

\S4method{-}{BFprobability,numeric}(e1, e2)

\S4method{[}{BFprobability,index,missing,missing}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{e1}{BFprobability object}

\item{e2}{new normalization constant}

\item{x}{BFprobability object}

\item{i}{indices indicating elements to extract}

\item{j}{unused for BFprobability objects}

\item{...}{further arguments passed to related methods}

\item{drop}{unused}
}
\description{
The \code{BFprobability} class is a general S4 class for representing models model comparison via prior or posterior probabilities.
}
\details{
\describe{
  The \code{BFprobability} class has the following slots defined:
   \item{odds}{A BFodds object containing the models from which to compute the probabilities}
   \item{normalize}{the sum of the probabilities of all models (will often be 1.0)}
   \item{version}{character string giving the version and revision number of the package that the model was created in}
   }
}

