% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aaGenerics.R, R/methods-BFodds.R
\docType{methods}
\name{extractOdds}
\alias{extractOdds}
\alias{extractOdds,BFodds-method}
\title{Extract the odds from an object}
\usage{
extractOdds(x, logodds = FALSE, onlyodds = FALSE)

\S4method{extractOdds}{BFodds}(x, logodds = FALSE, onlyodds = FALSE)
}
\arguments{
\item{x}{object from which to extract}

\item{logodds}{return the logarithm}

\item{onlyodds}{return a vector of only the odds}
}
\value{
Returns an object containing odds extracted from the object
}
\description{
Extract the odds from an object
}

