% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaGenerics.R, R/methods-BFprobability.R
\docType{methods}
\name{extractProbabilities}
\alias{extractProbabilities}
\alias{extractProbabilities,BFprobability-method}
\title{Extract the probabilities from an object}
\usage{
extractProbabilities(x, logprobs = FALSE, onlyprobs = FALSE)

\S4method{extractProbabilities}{BFprobability}(x, logprobs = FALSE,
  onlyprobs = FALSE)
}
\arguments{
\item{x}{object from which to extract}

\item{logprobs}{return the logarithm}

\item{onlyprobs}{return a vector of only the probabilities}
}
\value{
Returns an object containing probabilities extracted from the object
}
\description{
Extract the probabilities from an object
}
