% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaGenerics.R
\name{\%same\%}
\alias{\%same\%}
\title{Compare two objects to see if they are the 'same', for some loose definition
of same}
\usage{
x \%same\% y
}
\arguments{
\item{x}{first object}

\item{y}{second object}
}
\value{
Returns \code{TRUE} or \code{FALSE}
}
\description{
Compare two objects to see if they are the 'same', for some loose definition
of same
}
