% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaGenerics.R
\name{\%termin\%}
\alias{\%termin\%}
\title{Find a model term in a vector of model terms}
\usage{
x \%termin\% table
}
\arguments{
\item{x}{the terms to be matched}

\item{table}{the terms to be matched against}
}
\value{
A logical vector of the same length as x, indicating if a 
match was located for each element of x.
}
\description{
Find a model term in a vector of model terms
}

