% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPriorOdds.R
\name{newPriorOdds}
\alias{newPriorOdds}
\title{Create prior odds from a Bayes factor object}
\usage{
newPriorOdds(bf, type = "equal")
}
\arguments{
\item{bf}{A BFBayesFactor object, eg, from an analysis}

\item{type}{The type of prior odds to create (by default "equal"; see details)}
}
\value{
A (prior) BFodds object, which can then be multiplied by the
BFBayesFactor object to obtain posterior odds.
}
\description{
Create a prior odds object from a Bayes factor object
}
\details{
This function takes a Bayes factor object and, using its structure and
specified type of prior odds, will create a prior odds object.

For now, the only type is "equal", which assigns equal prior odds to all
models.
}
\author{
Richard D. Morey (\email{richarddmorey@gmail.com})
}
\keyword{misc}
