% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aaGenerics.R, R/methods-BFodds.R
\docType{methods}
\name{priorLogodds<-}
\alias{priorLogodds<-}
\alias{priorLogodds<-,BFodds,numeric-method}
\title{Set prior log odds in an object}
\usage{
priorLogodds(object) <- value

\S4method{priorLogodds}{BFodds,numeric}(object) <- value
}
\arguments{
\item{object}{object in which to set log odds}

\item{value}{log odds}
}
\description{
Set prior log odds in an object
}

