% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aaGenerics.R, R/methods-BFodds.R
\docType{methods}
\name{priorOdds<-}
\alias{priorOdds<-}
\alias{priorOdds<-,BFodds,numeric-method}
\title{Set prior odds in an object}
\usage{
priorOdds(object) <- value

\S4method{priorOdds}{BFodds,numeric}(object) <- value
}
\arguments{
\item{object}{object in which to set odds}

\item{value}{odds}
}
\description{
Set prior odds in an object
}

