% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesFactorPCL-package.R
\docType{data}
\name{puzzles}
\alias{puzzles}
\title{Puzzle completion times from Hays (1994)}
\format{
A data frame with 48 observations on 3 variables.  \describe{
\item{RT}{Puzzle completion time, in minutes} \item{ID}{the
subject identifier} \item{shape}{shape of the puzzle (round or
square)} \item{color}{color content of the puzzle (monochromatic or
color)} }
}
\source{
Hays, W. L. (1994), Statistics (5th edition), Harcourt Brace, Fort
Worth, Texas
}
\description{
Puzzle completion time example data from Hays (1994).
}
\details{
Hays (1994; section 13.21, table 13.21.2, p. 570) describes a experiment
wherein 12 participants complete four puzzles each. Puzzles could be either
square or round, and either monochromatic or in color. Each participant
completed every combination of the two factors.
}
\examples{

data(puzzles)

## classical ANOVA
## Both color and shape are significant, interaction is not
classical <- aov(RT ~ shape*color + Error(ID/(shape*color)), data=puzzles)
summary(classical)

## Bayes Factor
## Best model is main effects model, no interaction
anovaBF(RT ~ shape*color + ID, data = puzzles, whichRandom = "ID", progress=FALSE)


}
\keyword{datasets}
