% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual.R
\name{BFManual}
\alias{BFManual}
\title{Opens the HTML manual for the BayesFactor package}
\usage{
BFManual()
}
\value{
\code{BFManual} returns \code{NULL} invisibly.
}
\description{
This function opens the HTML manual for the BayesFactor package in whatever
browser is configured.
}
\details{
This function opens the HTML manual for the BayesFactor package in whatever
browser is configured.
}
\author{
Richard D. Morey (\email{richarddmorey@gmail.com})
}
\keyword{misc}
